requirements.txt
amqp==2.3.2
autopep8==1.3.5
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.4
cachetools==3.0.0
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
CoffeeScript==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2
defusedxml==0.7.0rc1
Django==2.2.18
django-anymail==3.0
django-appconf==1.0.3
django-bootstrap4==0.0.7
django-celery-beat==1.5.0
django-celerybeat-status==0.0.10
django-datatables-view==1.16.0
django-dynamic-preferences==1.10.1
django-extensions==2.1.6
django-filter==2.0.0
django-imagekit==4.0.2
django-oidc-provider==0.7.0
django-phonenumber-field==2.0.0
django-simple-history==2.12.0
django-tables2==1.21.2
django-timezone-field==3.0
django-twilio==0.11.0
djangorestframework==3.11.2
docutils==0.14
flake8==3.5.0
future==0.17.1
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.19.0
idna==2.6
isort==4.3.4
itypes==1.1.0
jedi==0.12.0
Jinja2==2.11.3
kombu==4.2.1
Mako==1.0.7
markdown2==2.3.9
MarkupSafe==1.0
mccabe==0.6.1
model-mommy==1.6.0
oauth2client==4.1.3
oauthlib==3.1.0
parso==0.2.1
pbr==4.2.0
persisting-theory==0.2.1
phonenumberslite==8.9.7
pilkit==2.0
Pillow==7.2.0
Plim==0.9.12
psycopg2-binary==2.7.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.3.1
pycparser==2.18
pycryptodomex==3.7.0
PyDrive==1.3.1
PyExecJS==1.5.1
pyflakes==1.6.0
pyjwkest==1.4.0
PyJWT==1.6.4
pyScss==1.3.5
PySocks==1.6.8
python-crontab==2.3.6
python-dateutil==2.7.3
python-dotenv==0.8.2
python3-openid==3.2.0
pytz==2018.4
PyYAML==5.3.1
raven==6.9.0
requests==2.20.0
requests-oauthlib==1.3.0
rope-py3k==0.9.4.post1
rsa==4.1
rules==2.0.1
six==1.11.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.0
stylus==0.1.2
twilio==6.14.3
uritemplate==3.0.0
urllib3==1.24.2
vine==1.1.4
vobject==0.9.6.1
Werkzeug==0.15.3
yapf==0.22.0
